+++ Game/Menu/Jeu.py
@@ -1,0 +1,42 @@
+import pygame
+import pygame_menu
+from pygame_menu import Menu
+from ..Config.config import *
+
+class Jeu:
+    """ Simulacre de l'interface du jeu """
+
+    def __init__(self, jeu, *groupes):
+        self._fenetre = jeu.fenetre
+        jeu.fond = background
+
+        from itertools import cycle
+        couleurs = [(0, 48, i) for i in range(0, 256, 15)]
+        couleurs.extend(sorted(couleurs[1:-1], reverse=True))
+        self._couleurTexte = cycle(couleurs)
+
+        self._font = pygame.font.SysFont('Helvetica', 36, bold=True)
+        self.creerTexte()
+        self.rectTexte = self.texte.get_rect()
+        self.rectTexte.center = (LARGEUR_FENETRE / 2, HAUTEUR_FENETRE / 2)
+        # Création d'un event
+        self._CLIGNOTER = pygame.USEREVENT + 1
+        pygame.time.set_timer(self._CLIGNOTER, 80)
+
+    def creerTexte(self):
+        self.texte = self._font.render(
+            'LE JEU EST EN COURS D\'EXÉCUTION',
+            True,
+            next(self._couleurTexte)
+        )
+
+    def update(self, events):
+        self._fenetre.blit(self.texte, self.rectTexte)
+        for event in events:
+            if event.type == self._CLIGNOTER:
+                self.creerTexte()
+                break
+
+    def detruire(self):
+        pygame.time.set_timer(self._CLIGNOTER, 0)  # désactivation du timer
+
