+++ Game/Menu/SettingMenu.py
@@ -1,7 +1,6 @@
 import pygame
 from ..Config.config import *
 
-
 class SettingMenu:
 
     pass
